// src/app/editor/models/canvas-templates.model.ts
export enum CanvasTemplates {
  cell = 'cell',
  fourCells = 'fourCells',
  columnThreeRows = 'columnThreeRows',
  columnTwoRows = 'columnTwoRows',
  rowThreeColumns = 'rowThreeColumns',
  rowTwoColumns = 'rowTwoColumns',
  theater = 'theater',
  threeColumns = 'threeColumns',
  threeRows = 'threeRows',
  twoColumns = 'twoColumns',
  twoColumnsRow = 'twoColumnsRow',
  twoRows = 'twoRows',
}